Local support - landing page
This screen snapshot shows landing page for all issues relating to the
Research Data Discovery Service at UTAS. It is designed to contain all
key elements to provide a quick turn around in providing assistance and
in resumption of services. It is provided here as a demonstration of how
a wiki can be used to provide high quality support services to users of
service of all levels.